Safari-609.1.11.1/LayoutTests/http/tests/security/credentials-iframes-allowCrossOriginSubresourcesToAskForCredentials.html

<!DOCTYPE html><!-- webkit-test-runner [ allowCrossOriginSubresourcesToAskForCredentials=true ] -->
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    testRunner.setCanOpenWindows();
    internals.settings.setStorageBlockingPolicy('BlockThirdParty');
}

window.addEventListener("message", function (event) {
    alert("parent host: " + window.location.hostname + event.data);
    window.open("http://localhost:8000/security/resources/credentials-iframes-continued.html");
}, false);
</script>
</head>
<body>
<iframe src="resources/credentials-iframes-same-domain.html"></iframe>
</body>
</html>
